refactor ooxml translators; remove static functions
